Add `flatten` attribute to derive SerializeRow
This commit focuses on the `match_by_name` flavor. All the needed
structs/traits to make this work are inside the macro internals module
as to not increase the public API surface area until we feel more
certain about the long term design. Effort was done to not change any
of the existing API.